Documentations

+++ R/genomaps.R
@@ -103,29 +103,33 @@
 #' Create a margeno object
 #'
 #' @param genotype Matrix of genotypes, where rows represent samples and columns represent variants. Each value represents the number of alternative alleles.
-#' @param ploidy Numeric value of ploidy. *Warning* ploidy other than 2 can be used, but result interpretation is not guaranteed.
+#' @param ploidy Integer. *Warning* ploidy other than 2 can be used, but result interpretation is not guaranteed.
 #' @param sample.id Character or integer or numeric vector of unique sample IDs.
 #'   Or NULL (the default), in which case samples are numbered from the genotype columns.
 #' @param variant.id Integer vector of unique variant IDs.
 #'   Or NULL (the default), in which case variants are numbered from the genotype rows.
-#' @param position Integer vector of variant positions. Or NULL.
-#' @param chromosome Character or integer vector of chromosome IDs. Or NULL.
+#' @param position Integer vector of variant positions. Or NULL (the default).
+#' @param chromosome Character or integer vector of chromosome IDs. Or NULL (the default).
 #'
-#' @return A margeno object
+#' @return A margeno object. In addition to the inputs, it stores `allele_count`,
+#'   the number of alternative alleles observed at each site across all samples.
+#' @seealso [text_parser()] and [vcf_parser()], which build a margeno object
+#'   directly from a file, and [genomaps()] to pair it with a [marmaps()] object.
 #' @export
 #'
 #' @examples
-#' genotype <- matrix(c(1, 2, 0, 0, 1, 1, 2, 0, 1, 2, 2, 1), nrow = 4, ncol = 3)
-#' ploidy <- 2
+#' # the genotypes behind the 1001 genomes example object
+#' geno <- gm1001g$geno
+#'
 #' # only genotype and ploidy are required
-#' margeno(genotype, ploidy)
+#' margeno(geno$genotype, geno$ploidy)
 #'
 #' # sample.id, variant.id, position and chromosome are all optional
-#' margeno(genotype, ploidy,
-#'     sample.id = c("sample1", "sample2", "sample3"),
-#'     variant.id = 1:4,
-#'     position = as.integer(c(100, 200, 300, 400)), # position has to be integer
-#'     chromosome = c("chr1", "chr1", "chr1", "chr2")
+#' margeno(geno$genotype, geno$ploidy,
+#'     sample.id = geno$sample.id,
+#'     variant.id = geno$variant.id,
+#'     position = geno$position,
+#'     chromosome = geno$chromosome
 #' )
 margeno <- function(genotype, ploidy, sample.id = NULL, variant.id = NULL, position = NULL, chromosome = NULL) {
     # validate data class
@@ -134,7 +138,7 @@ margeno <- function(genotype, ploidy, sample.id = NULL, variant.id = NULL, posit
     stopifnot(class(position) == "integer" | is.null(position))
     stopifnot(class(chromosome) %in% c("character", "integer") | is.null(chromosome))
     stopifnot("matrix" %in% class(genotype))
-    stopifnot(class(ploidy) == "numeric")
+    stopifnot(class(ploidy) == "integer")
 
     # if not provided, fill in sample.id and variant.id
     if (is.null(sample.id)) sample.id <- seq_len(ncol(genotype))
@@ -166,35 +170,44 @@ margeno <- function(genotype, ploidy, sample.id = NULL, variant.id = NULL, posit
 
 #' Create a marmaps object
 #'
-#' @param lonlatdf A data frame with three columns: sample ID, longitude, and latitude.
-#' @param mapres Optional numeric value for the map resolution. If not provided (mapres = NULL), it will be automatically calculated.
-#' @param mapcrs A character string specifying the coordinate reference system (CRS) for the map.
+#' Rasterizes the sampling coordinates of a species onto a regular grid. The grid
+#' is the unit of area for every downstream MAR calculation, so the map's coordinate
+#' reference system should be an equal-area projection whenever the samples span a wide latitudinal
+#' range.
+#'
+#' @param lonlatdf A data frame with three columns: sample ID, longitude, and latitude. Column names are not checked.
+#' @param mapcrs A character string specifying the coordinate reference system (CRS) for the output map.
+#' @param mapres Optional numeric value for the output map resolution, in the units of `mapcrs`.
+#'   If not provided (mapres = NULL), it will be automatically calculated.
 #' @param incrs A character string specifying the CRS of the input coordinates in `lonlatdf`.
 #'   Defaults to "EPSG:4326" (WGS84 longitude/latitude). When `incrs` differs from `mapcrs`,
 #'   the coordinates are reprojected into `mapcrs` before the resolution, extent, and cell
 #'   assignment are computed, so those stay consistent with the map units.
+#'   Set both `incrs` and `mapcrs` to `""` for coordinates on an arbitrary plane,
+#'   such as those of a spatial simulation.
 #'
-#' @return A marmaps object.
+#' @return A marmaps object, holding the sample IDs, the coordinates in `mapcrs`,
+#'   the sample map raster (wrapped with `terra::wrap()` so that it can be saved,
+#'   unwrap it with `terra::unwrap()` before use), and the map cell each sample
+#'   falls in.
+#' @seealso [lonlat_parser()], which builds a marmaps object directly from a
+#'   file, and [genomaps()] to pair it with a [margeno()] object.
 #' @export
 #'
 #' @examples
-#' lonlatdf <- data.frame(
-#'     id = c("sample1", "sample2", "sample3"),
-#'     longitude = c(-122.1, -121.9, -122.0),
-#'     latitude = c(37.8, 37.7, 37.9),
-#'     stringsAsFactors = FALSE
-#' )
-#' # input is WGS84 lon/lat, reprojected to an equal-area map
-#' marmaps(lonlatdf, mapres = NULL, mapcrs = "EPSG:8857", incrs = "EPSG:4326")
+#' # coordinates of the 1001 genomes accessions, in WGS84 lon/lat
+#' lonlat.fn <- system.file("extdata", "1001g_lonlat.txt", package = "mar")
+#' lonlatdf <- read.delim(lonlat.fn)
+#' head(lonlatdf)
 #'
-#' # input coordinates already in the target projection (no reprojection)
-#' lonlatdf <- data.frame(
-#'     id = c("sample1", "sample2", "sample3"),
-#'     x = c(100, 200, 400),
-#'     y = c(37.8, 37.7, 37.9),
-#'     stringsAsFactors = FALSE
-#' )
-#' marmaps(lonlatdf, mapres = NULL, mapcrs = "EPSG:8859", incrs = "EPSG:8859")
+#' # reprojected onto an equal-area map, resolution chosen automatically
+#' marmaps(lonlatdf, mapcrs = "EPSG:8857", incrs = "EPSG:4326")
+#'
+#' # the same map at a coarser, fixed resolution of 200 km
+#' marmaps(lonlatdf, mapcrs = "EPSG:8857", mapres = 2e5)
+#'
+#' # coordinates already in the target projection are not reprojected
+#' marmaps(lonlatdf, mapcrs = "EPSG:4326", incrs = "EPSG:8857")
 marmaps <- function(lonlatdf, mapcrs, mapres = NULL, incrs = "EPSG:4326") {
     # unpack lonlatdf
     stopifnot(class(lonlatdf) == "data.frame" & ncol(lonlatdf) == 3)
@@ -255,24 +268,22 @@ marmaps <- function(lonlatdf, mapcrs, mapres = NULL, incrs = "EPSG:4326") {
 #' @param geno A margeno object.
 #' @param maps A marmaps object.
 #'
-#' @return A genomaps object.
+#' @return A genomaps object, a list of the `geno` and `maps` inputs.
+#' @seealso [margeno()] and [marmaps()] for the two components, and
+#'   [MARsampling()], [MARextinction()] and [MARtheory()] for using genomaps.
 #' @export
 #'
 #' @examples
-#' sample_id <- c("sample1", "sample2", "sample3")
-#' genotype <- matrix(c(1, 2, 0, 0, 1, 1, 2, 0, 1, 2, 2, 1), nrow = 4, ncol = 3)
-#' ploidy <- 2
-#' geno <- margeno(genotype, ploidy, sample.id = sample_id)
+#' # gm1001g is itself a genomaps object, rebuild it from its two components
+#' geno <- gm1001g$geno
+#' maps <- gm1001g$maps
+#' gm <- genomaps(geno, maps)
 #'
-#' lonlatdf <- data.frame(
-#'     id = sample_id,
-#'     longitude = c(-122.1, -121.9, -122.0),
-#'     latitude = c(37.8, 37.7, 37.9),
-#'     stringsAsFactors = FALSE
-#' )
-#' maps <- marmaps(lonlatdf, mapres = NULL, mapcrs = "EPSG:4326")
-#'
-#' genomaps(geno, maps)
+#' print(gm)
+#' # genomaps() will fail if sample names are not matching
+#' maps_bad <- gm1001g$maps
+#' maps_bad$sample.id <- 1:1004
+#' gm <- genomaps(geno, maps_bad)
 genomaps <- function(geno, maps) {
     # validate inputs
     stopifnot("geno must be a margeno object" = "margeno" %in% class(geno))
